group boxes for the normal functions and the extra function as well as the text box that displays the results from a calculation are not visible during run time until there is text inside the input boxes.

+++ Calculator/Calculate.cs
@@ -18,7 +18,7 @@ namespace Calculator
             return Convert.ToString(a + b);
         }
 
-        public string subtract(int a, int b)
+        public string subtract(double a, double b)
         {
             return Convert.ToString(a - b);
         }
